requirements.txt
flask~=2.2.5
pymongo~=4.7.3
boto3~=1.34.120
pydantic~=2.7.4
jsonschema~=4.22.0
django-scheduler~=0.10.1
python-dotenv~=1.0.1
flask-cors
matplotlib~=3.9.0
Cartopy~=0.23.0
requests~=2.32.2
pandas
pyarrow
apscheduler
cryptography~=43.0.3
PyJWT~=2.10.1
tf-keras
tensorflow
transformers~=4.46.3
spacy~=3.8.2
openai~=1.56.1
firebase-admin
flask_cors
firebase-admin